refactor post-processing logic to postprocessing.go, and optimize postprocessing code
